requirements.txt
defusedxml==0.7.1
et_xmlfile==2.0.0
Faker==37.5.3
greenlet==3.2.4
numpy==2.2.6
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
python-calamine==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxlsb==1.0.10
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
tzdata==2025.2
xlrd==2.0.2
xlsxwriter==3.2.5
